+++ basic_drift.sh
@@ -1,0 +1,5 @@
+#! /bin/bash
+set -eou pipefail
+bash2gitlab detect-drift --out test/scenario2/out
+export NO_COLOR=NO_COLOR
+bash2gitlab detect-drift --out test/scenario2/out